Fix speecht5 dtype mismatch for bfloat16 models
decoder_input_values was created as float32 but ClarttsAr model weights
are bfloat16, causing a mat1/mat2 dtype mismatch at runtime.

+++ speecht5/pytorch/loader.py
@@ -153,7 +153,9 @@ class ModelLoader(ForgeModel):
         # Create tokenized inputs
         model_inputs = self.processor(text=self.sample_text, return_tensors="pt")
 
-        decoder_input_values = torch.zeros((1, 1, self.model.config.num_mel_bins))
+        decoder_input_values = torch.zeros(
+            (1, 1, self.model.config.num_mel_bins), dtype=self.model.dtype
+        )
 
         # Return inputs as a dictionary
         inputs = {
